Ionic Grocery App Screenshots
This repository is filled with screen shots of ionic code that is for a grocery app. The code screenshots are the coding for each tab that is displayed for the first two tabs. Tab1 is a grocery tab that displays the list of food and the quantity for that particular food. There is also coding done for interaction with the user. When the user swipes on the listed item they are able to delete the item from the list. The tab2 is an about me tab that displays my name and my student ID for Maryville University. The other three screenshots are of the user interface of the app on both iOS and Android. The first one shows the grocery tab, the second shows the grocery tab with the slide functionality, and the last one shows the about me tab.